Move genValidSelector util func to common module
